+++ projects/igniteui-angular-i18n/src/i18n/PL/carousel-resources.ts
@@ -1,0 +1,14 @@
+import { ICarouselResourceStrings } from 'igniteui-angular';
+
+// exported below as re-cast to create declaration type with expanded properties
+const CarouselResourceStringsPL_: ExpandRequire<ICarouselResourceStrings> = {
+    igx_carousel_of: 'z',
+    igx_carousel_slide: 'slajd',
+    igx_carousel_previous_slide: 'poprzedni slajd',
+    igx_carousel_next_slide: 'następny slajd'
+};
+
+/**
+ * Polish resource strings for IgxCarousel
+ */
+export const CarouselResourceStringsPL = CarouselResourceStringsPL_ as ExpandRequire<ICarouselResourceStrings>;