+++ projects/igniteui-angular-i18n/src/i18n/TR/tree-resources.ts
@@ -1,0 +1,12 @@
+import { ITreeResourceStrings } from 'igniteui-angular';
+
+// exported below as re-cast to create declaration type with expanded properties
+const TreeResourceStringsTR_: ExpandRequire<ITreeResourceStrings> = {
+    igx_expand: 'Genişlet',
+    igx_collapse: 'Daralt'
+};
+
+/**
+ * Turkish resource strings for IgxTree
+ */
+export const TreeResourceStringsTR = TreeResourceStringsTR_ as ExpandRequire<ITreeResourceStrings>;