+++ projects/igniteui-angular/src/lib/test-utils/grid-samples.spec.ts
@@ -2,7 +2,6 @@ import { Component, TemplateRef, ViewChild, Input, AfterViewInit, ChangeDetector
 import { IgxGridCellComponent } from '../grids/cell.component';
 import { IgxDateSummaryOperand, IgxNumberSummaryOperand, IgxSummaryResult } from '../grids/summaries/grid-summary';
 import { IGridCellEventArgs, IGridEditEventArgs } from '../grids/common/events';
-import { IgxGridTransaction } from '../grids/grid-base.directive';
 import {
     BasicGridComponent, BasicGridSearchComponent, GridAutoGenerateComponent,
     GridNxMComponent, GridWithSizeComponent, PagingComponent
@@ -11,7 +10,6 @@ import { IGridSelection } from './grid-interfaces.spec';
 import { SampleTestData, DataParent } from './sample-test-data.spec';
 import { ColumnDefinitions, GridTemplateStrings, EventSubscriptions } from './template-strings.spec';
 import { IgxColumnComponent } from '../grids/columns/column.component';
-import { IgxTransactionService } from '../services/public_api';
 import { IgxFilteringOperand, IgxNumberFilteringOperand } from '../data-operations/filtering-condition';
 import { ExpressionUI } from '../grids/filtering/grid-filtering.service';
 import { IFilteringExpressionsTree, FilteringExpressionsTree } from '../data-operations/filtering-expressions-tree';
@@ -762,7 +760,7 @@ export class GridHireDateComponent extends BasicGridComponent {
 @Component({
     template: `<div style="margin: 50px;">
             ${GridTemplateStrings.declareGrid(
-        `[height]="height" [width]="width" [rowSelection]="rowSelection" [autoGenerate]="autoGenerate"`,
+        '[height]="height" [width]="width" [rowSelection]="rowSelection" [autoGenerate]="autoGenerate"',
         EventSubscriptions.columnMovingStart + EventSubscriptions.columnMoving + EventSubscriptions.columnMovingEnd,
         ColumnDefinitions.movableColumns)}</div>`
 })
@@ -811,7 +809,7 @@ export class MovableColumnsComponent extends BasicGridComponent {
 }
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="300px" width="500px"`, '', ColumnDefinitions.movableColumns)}`
+    template: GridTemplateStrings.declareGrid(`height="300px" width="500px"`, '', ColumnDefinitions.movableColumns)
 })
 export class MovableTemplatedColumnsComponent extends BasicGridComponent {
     public data = SampleTestData.personIDNameRegionData();
@@ -821,8 +819,8 @@ export class MovableTemplatedColumnsComponent extends BasicGridComponent {
 }
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="300px" width="500px" [autoGenerate]="autoGenerate" [paging]="paging"`,
-        EventSubscriptions.columnInit, '')}`
+    template: GridTemplateStrings.declareGrid(`height="300px" width="500px" [autoGenerate]="autoGenerate" [paging]="paging"`,
+        EventSubscriptions.columnInit, '')
 })
 export class MovableColumnsLargeComponent extends GridAutoGenerateComponent {
 
@@ -845,7 +843,7 @@ export class MovableColumnsLargeComponent extends GridAutoGenerateComponent {
 }
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="800px"`, '', ColumnDefinitions.multiColHeadersColumns)}`
+    template: GridTemplateStrings.declareGrid(`height="800px"`, '', ColumnDefinitions.multiColHeadersColumns)
 })
 export class MultiColumnHeadersComponent extends BasicGridComponent {
     public data = SampleTestData.contactInfoDataFull();
@@ -853,8 +851,8 @@ export class MultiColumnHeadersComponent extends BasicGridComponent {
 }
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="800px"  width="500px"`, '',
-        ColumnDefinitions.multiColHeadersWithGroupingColumns)}`
+    template: GridTemplateStrings.declareGrid(`height="800px"  width="500px"`, '',
+        ColumnDefinitions.multiColHeadersWithGroupingColumns)
 })
 export class MultiColumnHeadersWithGroupingComponent extends BasicGridComponent {
     public data = SampleTestData.contactInfoDataFull();
@@ -863,7 +861,7 @@ export class MultiColumnHeadersWithGroupingComponent extends BasicGridComponent 
 
 
 @Component({
-    template: `${GridTemplateStrings.declareBasicGridWithColumns(ColumnDefinitions.nameAvatar)}`
+    template: GridTemplateStrings.declareBasicGridWithColumns(ColumnDefinitions.nameAvatar)
 })
 export class GridWithAvatarComponent extends GridWithSizeComponent {
     public data = SampleTestData.personAvatarData();
@@ -872,8 +870,8 @@ export class GridWithAvatarComponent extends GridWithSizeComponent {
 
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="1000px"  width="900px" [primaryKey]="'ID'"`, '',
-        ColumnDefinitions.summariesGroupByColumns, '', '<igx-paginator *ngIf="paging"></igx-paginator>')}`
+    template: GridTemplateStrings.declareGrid(`height="1000px"  width="900px" primaryKey="ID"`, '',
+        ColumnDefinitions.summariesGroupByColumns, '', '<igx-paginator *ngIf="paging"></igx-paginator>')
 })
 export class SummariesGroupByComponent extends BasicGridComponent {
     public data = SampleTestData.employeeGroupByData();
@@ -884,9 +882,8 @@ export class SummariesGroupByComponent extends BasicGridComponent {
 }
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="600px"  width="900px" [primaryKey]="'ID'"`, '',
-        ColumnDefinitions.summariesGroupByTansColumns)}`,
-    providers: [{ provide: IgxGridTransaction, useClass: IgxTransactionService }]
+    template: GridTemplateStrings.declareGrid(`height="600px"  width="900px" [batchEditing]="true" primaryKey="ID"`, '',
+    ColumnDefinitions.summariesGroupByTansColumns)
 })
 export class SummariesGroupByTransactionsComponent extends BasicGridComponent {
     public data = SampleTestData.employeeGroupByData();
@@ -896,8 +893,8 @@ export class SummariesGroupByTransactionsComponent extends BasicGridComponent {
 }
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="800px"  width="400px" [primaryKey]="'ID'"`, '',
-        ColumnDefinitions.summariesGroupByColumns)}`
+    template: GridTemplateStrings.declareGrid('height="800px"  width="400px" primaryKey="ID"', '',
+        ColumnDefinitions.summariesGroupByColumns)
 })
 export class SummariesGroupByWithScrollsComponent extends BasicGridComponent {
     public data = SampleTestData.employeeGroupByData();
@@ -1004,8 +1001,8 @@ export class GridWithPrimaryKeyComponent extends BasicGridSearchComponent {
 }
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="300px"  width="600px" [primaryKey]="'ID'"`, '',
-        ColumnDefinitions.selectionWithScrollsColumns, '', '<igx-paginator *ngIf="paging"></igx-paginator>')}`,
+    template: GridTemplateStrings.declareGrid(`height="300px"  width="600px" primaryKey="ID"`, '',
+        ColumnDefinitions.selectionWithScrollsColumns, '', '<igx-paginator *ngIf="paging"></igx-paginator>'),
 })
 export class SelectionWithScrollsComponent extends BasicGridComponent {
     public data = SampleTestData.employeeGroupByData();
@@ -1013,24 +1010,23 @@ export class SelectionWithScrollsComponent extends BasicGridComponent {
 }
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="300px"  width="600px" [primaryKey]="'ID'" cellSelection="none"`, '',
-        ColumnDefinitions.selectionWithScrollsColumns)}`,
+    template: GridTemplateStrings.declareGrid(`height="300px"  width="600px" primaryKey="ID" cellSelection="none"`, '',
+        ColumnDefinitions.selectionWithScrollsColumns),
 })
 export class CellSelectionNoneComponent extends BasicGridComponent {
     public data = SampleTestData.employeeGroupByData();
 }
 
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="300px"  width="600px" [primaryKey]="'ID'" cellSelection="single"`, '',
-        ColumnDefinitions.selectionWithScrollsColumns)}`,
+    template: GridTemplateStrings.declareGrid(`height="300px"  width="600px" primaryKey="ID" cellSelection="single"`, '',
+        ColumnDefinitions.selectionWithScrollsColumns),
 })
 export class CellSelectionSingleComponent extends BasicGridComponent {
     public data = SampleTestData.employeeGroupByData();
 }
 @Component({
-    template: `${GridTemplateStrings.declareGrid(`height="300px"  width="600px" [primaryKey]="'ID'"`, '',
-        ColumnDefinitions.selectionWithScrollsColumns)}`,
-    providers: [{ provide: IgxGridTransaction, useClass: IgxTransactionService }]
+    template: GridTemplateStrings.declareGrid(`height="300px"  width="600px" [batchEditing]="true" primaryKey="ID"`, '',
+        ColumnDefinitions.selectionWithScrollsColumns)
 })
 export class SelectionWithTransactionsComponent extends BasicGridComponent {
     public data = SampleTestData.employeeGroupByData();
@@ -1752,15 +1748,14 @@ export class IgxGridEmptyRowEditTemplateComponent extends BasicGridComponent {
 
 @Component({
     template: `
-    <igx-grid #grid [data]="data" [primaryKey]="'ProductID'" width="900px" height="900px" [rowEditable]="true" >
+    <igx-grid #grid [data]="data" [batchEditing]="true" [primaryKey]="'ProductID'" width="900px" height="900px" [rowEditable]="true" >
         <igx-column field="ProductID" header="Product ID" width="150px"></igx-column>
         <igx-column field="ProductName" header="Product Name" [dataType]="'string'" width="200px"></igx-column>
         <igx-column field="InStock" header="In Stock" [dataType]="'boolean'" width="100px"></igx-column>
         <igx-column field="UnitsInStock" header="Units in Stock" [dataType]="'number'" width="150px"></igx-column>
         <igx-column field="OrderDate" header="Order Date" [dataType]="'date'" width="200px"></igx-column>
         <igx-paginator *ngIf='paging' [perPage]="7"></igx-paginator>
-    </igx-grid>`,
-    providers: [{ provide: IgxGridTransaction, useClass: IgxTransactionService }],
+    </igx-grid>`
 })
 export class IgxGridRowEditingTransactionComponent extends BasicGridComponent {
     public data = SampleTestData.foodProductData();
@@ -1770,14 +1765,14 @@ export class IgxGridRowEditingTransactionComponent extends BasicGridComponent {
 
 @Component({
     template: `
-    <igx-grid #grid [data]="data" [primaryKey]="'ProductID'" width="900px" height="900px" [paging]="paging" [perPage]="7">
+    <igx-grid #grid [data]="data" [batchEditing]="true"
+        [primaryKey]="'ProductID'" width="900px" height="900px" [paging]="paging" [perPage]="7">
         <igx-column field="ProductID" header="Product ID" width="150px"></igx-column>
         <igx-column field="ProductName" header="Product Name" [dataType]="'string'" width="200px"></igx-column>
         <igx-column field="InStock" header="In Stock" [dataType]="'boolean'" width="100px"></igx-column>
         <igx-column field="UnitsInStock" header="Units in Stock" [dataType]="'currency'" width="150px"></igx-column>
         <igx-column field="OrderDate" header="Order Date" [dataType]="'date'" width="200px"></igx-column>
-    </igx-grid>`,
-    providers: [{ provide: IgxGridTransaction, useClass: IgxTransactionService }],
+    </igx-grid>`
 })
 export class IgxGridCurrencyColumnComponent extends BasicGridComponent {
     public data = SampleTestData.foodProductData();
@@ -1786,14 +1781,13 @@ export class IgxGridCurrencyColumnComponent extends BasicGridComponent {
 
 @Component({
     template: `
-    <igx-grid #grid [data]="data" [primaryKey]="'ProductID'" width="900px" height="900px">
+    <igx-grid #grid [data]="data" [batchEditing]="true" [primaryKey]="'ProductID'" width="900px" height="900px">
         <igx-column field="ProductID" header="Product ID" width="150px"></igx-column>
         <igx-column field="ProductName" header="Product Name" [dataType]="'string'" width="200px"></igx-column>
         <igx-column field="InStock" header="In Stock" [dataType]="'boolean'" width="100px"></igx-column>
         <igx-column field="UnitsInStock" header="Units in Stock" [dataType]="'currency'" width="150px"></igx-column>
         <igx-column field="Discount" header="Order Date" [dataType]="'percent'" width="200px"></igx-column>
-    </igx-grid>`,
-    providers: [{ provide: IgxGridTransaction, useClass: IgxTransactionService }],
+    </igx-grid>`
 })
 export class IgxGridPercentColumnComponent extends BasicGridComponent {
     public data = SampleTestData.foodPercentProductData();
@@ -1801,15 +1795,15 @@ export class IgxGridPercentColumnComponent extends BasicGridComponent {
 
 @Component({
     template: `
-    <igx-grid #grid [data]="data" [primaryKey]="'ProductID'" width="900px" height="900px" [paging]="paging" [perPage]="7">
+    <igx-grid #grid [data]="data" [batchEditing]="true"
+    [primaryKey]="'ProductID'" width="900px" height="900px" [paging]="paging" [perPage]="7">
         <igx-column field="ProductID" header="Product ID" width="150px"></igx-column>
         <igx-column field="ProductName" header="Product Name" [dataType]="'string'" width="200px"></igx-column>
         <igx-column field="OrderDate" header="Order Date" [dataType]="'dateTime'" width="250px"></igx-column>
         <igx-column field="ReceiveTime" header="Receive Time" [dataType]="'time'" width="200px"></igx-column>
         <igx-column field="InStock" header="In Stock" [dataType]="'boolean'" width="100px"></igx-column>
         <igx-column field="UnitsInStock" header="Units in Stock" [dataType]="'currency'" width="150px"></igx-column>
-    </igx-grid>`,
-    providers: [{ provide: IgxGridTransaction, useClass: IgxTransactionService }],
+    </igx-grid>`
 })
 export class IgxGridDateTimeColumnComponent extends BasicGridComponent {
     public data = SampleTestData.foodProductDateTimeData();